refactor: Move groups related endpoints into GroupController
