+++ src/main/java/org/isf/config/SecurityConfig.java
@@ -251,15 +251,15 @@ public class SecurityConfig {
 			.requestMatchers(HttpMethod.PUT, "/permissions/**").hasAuthority("permissions.update")
 			.requestMatchers(HttpMethod.DELETE, "/permissions/**").hasAuthority("permissions.delete")
 			// grouppermission
-			.requestMatchers(HttpMethod.POST, "/users/groups/{group_code}/permissions/**").hasAuthority("grouppermission.create")
-			.requestMatchers(HttpMethod.GET, "/users/groups/{group_code}/permissions/**").hasAuthority("grouppermission.read")
-			.requestMatchers(HttpMethod.PUT, "/users/groups/{group_code}/permissions/**").hasAuthority("grouppermission.update")
-			.requestMatchers(HttpMethod.DELETE, "/users/groups/{group_code}/permissions/**").hasAuthority("grouppermission.delete")
+			.requestMatchers(HttpMethod.POST, "/groups/{group_code}/permissions/**").hasAuthority("grouppermission.create")
+			.requestMatchers(HttpMethod.GET, "/groups/{group_code}/permissions/**").hasAuthority("grouppermission.read")
+			.requestMatchers(HttpMethod.PUT, "/groups/{group_code}/permissions/**").hasAuthority("grouppermission.update")
+			.requestMatchers(HttpMethod.DELETE, "/groups/{group_code}/permissions/**").hasAuthority("grouppermission.delete")
 			// usergroups
-			.requestMatchers(HttpMethod.POST, "/users/groups/**").hasAuthority("usergroups.create")
-			.requestMatchers(HttpMethod.GET, "/users/groups/**").hasAuthority("usergroups.read")
-			.requestMatchers(HttpMethod.PUT, "/users/groups/**").hasAuthority("usergroups.update")
-			.requestMatchers(HttpMethod.DELETE, "/users/groups/**").hasAuthority("usergroups.delete")
+			.requestMatchers(HttpMethod.POST, "/groups/**").hasAuthority("usergroups.create")
+			.requestMatchers(HttpMethod.GET, "/groups/**").hasAuthority("usergroups.read")
+			.requestMatchers(HttpMethod.PUT, "/groups/**").hasAuthority("usergroups.update")
+			.requestMatchers(HttpMethod.DELETE, "/groups/**").hasAuthority("usergroups.delete")
 			// user
 			.requestMatchers(HttpMethod.POST, "/users").hasAuthority("users.create")
 			.requestMatchers(HttpMethod.GET, "/users/**").hasAuthority("users.read")
